Refactor grade-related logic via code extraction
